Fixed variance and test_functions to work as intended
Also added two import lines in test_functions to allow for refreshing of variance and ml_matrix_composition files after reloading (lines 12-13). Before, if I changed variance and reran test_variance() it would use the old version. With this added in, it uses the new version. I'm currently not sure why this works this way